Correct prototype of OGR_ST_SetParamDbl() in .cpp to match with the one in .h
git-svn-id: https://svn.osgeo.org/gdal/trunk/gdal@18572 f0d54148-0727-0410-94bb-9a71ac55c965

+++ ogr/ogrfeaturestyle.cpp
@@ -2376,7 +2376,7 @@ void OGR_ST_SetParamNum( OGRStyleToolH hST, int eParam, int nValue )
  *
  */
 
-void OGR_ST_SetParamDbl( OGRStyleToolH hST, int eParam, int dfValue )
+void OGR_ST_SetParamDbl( OGRStyleToolH hST, int eParam, double dfValue )
 {
     VALIDATE_POINTER0( hST, "OGR_ST_SetParamDbl" );
